+++ NiceDay/src/world/block/Block.h
@@ -109,10 +109,10 @@ protected:
 	bool m_has_big_texture;
 
 	//how much light will be consumed by this block 
-	//1		-> consumes all light
-	//0.5	-> consumes half of light passing through
-	//0		-> consumes no light
-	float m_opacity;
+	//32	-> consumes all light
+	//1		-> consumes one light level
+	//0		-> consumes no light (you probably dont want that)
+	uint8_t m_opacity;
 
 	int m_block_connect_group;
 	bool isInGroup(World* w, int x, int y, int group) const;
@@ -123,7 +123,7 @@ public:
 	virtual ~Block();
 	inline int getID() const { return m_id; };
 	inline int getConnectGroup() const { return m_block_connect_group; }
-	virtual float getOpacity(const BlockStruct&) const;
+	virtual uint8_t getOpacity(const BlockStruct&) const;
 
 	inline bool isInConnectGroup(int groups) const { return (groups & m_block_connect_group) != 0; }//they have group in common
